# bpo-39151: Simplify dfs in the assembler