--- a/Python/compile.c
+++ b/Python/compile.c
@@ -5447,7 +5447,7 @@ dfs(struct compiler *c, basicblock *b, struct assembler *a, int end)
 {
 
     /* There is no real depth-first-search to do here because all the
-     * blocks are emmited in topological order already, so we just need to
+     * blocks are emitted in topological order already, so we just need to
      * follow the b_next pointers and place them in a->a_reverse_postorder in
      * reverse order and make sure that the first one starts at 0. */