This simplification has the advantage that we don't stack overflow if the amount of jumps in the blocks is insane (maybe some autogenerated code).

The current code *already* relies on the topological order of the blocks, this PR just continues with that assumption to further simplify.

<!-- issue-number: [bpo-39151](https://bugs.python.org/issue39151) -->
https://bugs.python.org/issue39151
<!-- /issue-number -->