@pablogsal 2020-01-03
As a bonus, this does hilariously jumpy code compilation faster:

## master
```
./python.exe -m pyperf timeit 'code = ("\n".join("  "*i + "if x:\n" + "  "*i + "  ...\n" + "  "*i + "else:" for i in range(30)) + "\n" + "  "*30 + "...")' 'compile(code, "<dummy>", "exec")'
.....................
Mean +- std dev: 785 us +- 16 us
```

## patched
```
./python.exe -m pyperf timeit 'code = ("\n".join("  "*i + "if x:\n" + "  "*i + "  ...\n" + "  "*i + "else:" for i in range(30)) + "\n" + "  "*30 + "...")' 'compile(code, "<dummy>", "exec")'
.....................
Mean +- std dev: 734 us +- 9 us
```